remove namedtuples out of cuthbertlib
