+++ cuthbertlib/kalman/__init__.py
@@ -1,6 +1,3 @@
-from cuthbertlib.kalman.filtering import KalmanState
-from cuthbertlib.kalman.filtering import KalmanFilterInfo
 from cuthbertlib.kalman.filtering import predict
 from cuthbertlib.kalman.filtering import update as filter_update
-from cuthbertlib.kalman.smoothing import KalmanSmootherInfo
 from cuthbertlib.kalman.smoothing import update as smoother_update
